+++ backend/ops_api/ops/urls.py
@@ -1,6 +1,6 @@
-from ops.can.models import BudgetLineItem
-from ops.can.models import CAN
-from ops.can.models import CANFiscalYear
+from ops.models.cans import BudgetLineItem
+from ops.models.cans import CAN
+from ops.models.cans import CANFiscalYear
 from ops.portfolio.models import Division
 from ops.portfolio.models import Portfolio
 from ops.portfolio.models import PortfolioStatus
