+++ backend/ops_api/tests/ops/can/test_agreement.py
@@ -1,4 +1,4 @@
-from ops.can.models import Agreement
+from ops.models.cans import Agreement
 import pytest
