+++ backend/ops_api/tests/ops/can/test_agreement_type.py
@@ -1,4 +1,4 @@
-from ops.can.models import AgreementType
+from ops.models.cans import AgreementType
 import pytest
