+++ backend/ops_api/tests/ops/can/test_budget_line_item.py
@@ -1,4 +1,4 @@
-from ops.can.models import BudgetLineItem
+from ops.models.cans import BudgetLineItem
 import pytest
