+++ backend/ops_api/tests/ops/can/test_budget_line_item_status.py
@@ -1,4 +1,4 @@
-from ops.can.models import BudgetLineItemStatus
+from ops.models.cans import BudgetLineItemStatus
 import pytest
