+++ backend/ops_api/tests/ops/can/test_can.py
@@ -1,4 +1,4 @@
-from ops.can.models import CAN
+from ops.models.cans import CAN
 import pytest
