+++ backend/ops_api/tests/ops/can/test_can_fiscal_year_carry_over.py
@@ -1,4 +1,4 @@
-from ops.can.models import CANFiscalYearCarryOver
+from ops.models.cans import CANFiscalYearCarryOver
 import pytest
