+++ backend/ops_api/tests/ops/can/test_funding_partner.py
@@ -1,4 +1,4 @@
-from ops.can.models import FundingPartner
+from ops.models.cans import FundingPartner
 import pytest
