+++ backend/ops_api/tests/ops/can/test_funding_source.py
@@ -1,4 +1,4 @@
-from ops.can.models import FundingSource
+from ops.models.cans import FundingSource
 import pytest
