+++ backend/ops_api/tests/ops/conftest.py
@@ -1,14 +1,14 @@
 from datetime import datetime
 
 from ops import create_app
-from ops.can.models import Agreement
-from ops.can.models import BudgetLineItem
-from ops.can.models import BudgetLineItemStatus
-from ops.can.models import CAN
-from ops.can.models import CANFiscalYear
-from ops.can.models import CANFiscalYearCarryOver
-from ops.can.models import FundingPartner
-from ops.can.models import FundingSource
+from ops.models.cans import Agreement
+from ops.models.cans import BudgetLineItem
+from ops.models.cans import BudgetLineItemStatus
+from ops.models.cans import CAN
+from ops.models.cans import CANFiscalYear
+from ops.models.cans import CANFiscalYearCarryOver
+from ops.models.cans import FundingPartner
+from ops.models.cans import FundingSource
 from ops.portfolio.models import Division
 from ops.portfolio.models import Portfolio
 from ops.portfolio.models import PortfolioDescriptionText
